Conflicts:
	inc/sitemaps/class-author-sitemap-provider.php

+++ admin/ajax.php
@@ -92,23 +92,35 @@ function wpseo_kill_blocking_files() {
 
 	check_ajax_referer( 'wpseo-blocking-files' );
 
-	$message = 'There were no files to delete.';
+	$message = 'success';
+	$errors  = array();
+
+	// Todo: Use WP_Filesystem, but not so easy to use in AJAX with credentials form still internal.
 	$options = get_option( 'wpseo' );
 	if ( is_array( $options['blocking_files'] ) && $options['blocking_files'] !== array() ) {
-		$message       = 'success';
-		$files_removed = 0;
-		foreach ( $options['blocking_files'] as $k => $file ) {
-			if ( ! @unlink( $file ) ) {
-				$message = __( 'Some files could not be removed. Please remove them via FTP.', 'wordpress-seo' );
+		foreach ( $options['blocking_files'] as $file ) {
+			if ( is_file( $file ) ) {
+				if ( ! @unlink( $file ) ) {
+					$errors[] = __(
+						sprintf( 'The file "%s" could not be removed. Please remove it via FTP.', $file ),
+						'wordpress-seo'
+					);
+				}
 			}
-			else {
-				unset( $options['blocking_files'][ $k ] );
-				$files_removed ++;
+
+			if ( is_dir( $file ) ) {
+				if ( ! @ rmdir( $file ) ) {
+					$errors[] = __(
+						sprintf( 'The directory "%s" could not be removed. Please remove it via FTP.', $file ),
+						'wordpress-seo'
+					);
+				}
 			}
 		}
-		if ( $files_removed > 0 ) {
-			update_option( 'wpseo', $options );
-		}
+	}
+
+	if ( $errors ) {
+		$message = implode( '<br />', $errors );
 	}
 
 	die( $message );
@@ -359,6 +371,39 @@ function ajax_get_term_keyword_usage() {
 }
 
 add_action( 'wp_ajax_get_term_keyword_usage',  'ajax_get_term_keyword_usage' );
+
+/**
+ * Removes stopword from the sample permalink that is generated in an AJAX request
+ *
+ * @param array  $permalink The permalink generated for this post by WordPress.
+ * @param int    $post_ID The ID of the post.
+ * @param string $title The title for the post that the user used.
+ * @param string $name The name for the post that the user used.
+ *
+ * @return array
+ */
+function wpseo_remove_stopwords_sample_permalink( $permalink, $post_ID, $title, $name ) {
+	WPSEO_Options::get_instance();
+	$options = WPSEO_Options::get_options( array( 'wpseo_permalinks' ) );
+	if ( $options['cleanslugs'] !== true ) {
+		return $permalink;
+	}
+
+	/*
+	 * If the name is empty and the title is not, WordPress will generate a slug. In that case we want to remove stop
+	 * words from the slug.
+	 */
+	if ( empty( $name ) && ! empty( $title ) ) {
+		$stop_words = new WPSEO_Admin_Stop_Words();
+
+		// The second element is the slug.
+		$permalink[1] = $stop_words->remove_in( $permalink[1] );
+	}
+
+	return $permalink;
+}
+
+add_action( 'get_sample_permalink', 'wpseo_remove_stopwords_sample_permalink', 10, 4 );
 
 // Crawl Issue Manager AJAX hooks.
 new WPSEO_GSC_Ajax;
